requirements.txt
ImageHash>=4.3.1
numpy>=1.26.0
Pillow>=10.0.0
streamlit>=1.28.0
opencv-python>=4.8.0
requests>=2.28.0
